Moved SENTINEL_INDEX and SENTINEL to tree.ts
