+++ src/page/tree/createNewPage.ts
@@ -1,7 +1,6 @@
 import { OnenotePage } from "@microsoft/microsoft-graph-types";
 import { Buffer, Color, Node, PageContent } from "../model";
-import { SENTINEL_INDEX } from "../reducer";
-import { getLineStarts, getNewlineFormat } from "./tree";
+import { getLineStarts, getNewlineFormat, SENTINEL_INDEX } from "./tree";
 
 /**
  * Creates a new page, and its associated piece table.
